+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/attendeeRepository/attendeeRepositoryImpls/AttendeeRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.realityexpander.tasky.agenda_feature.data.repositories.attendeeReposi
 import com.realityexpander.tasky.agenda_feature.data.repositories.attendeeRepository.remote.IAttendeeApi
 import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
 import com.realityexpander.tasky.agenda_feature.domain.Attendee
-import com.realityexpander.tasky.agenda_feature.domain.ResultUiText
+import com.realityexpander.tasky.core.presentation.util.ResultUiText
 import com.realityexpander.tasky.auth_feature.domain.validation.ValidateEmail
 import com.realityexpander.tasky.core.presentation.util.UiText
 import com.realityexpander.tasky.core.util.Email
