+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/eventRepository/remote/eventApi/DTOs/EventDTO.kt
@@ -57,7 +57,6 @@ abstract class EventDTO : AbstractAgendaItem(), UsesEpochMilli {
     ) : EventDTO()
 
     @Serializable
-    @OptIn(ExperimentalSerializationApi::class) // for JsonNames
     data class Response constructor(  // RESPONSE only
         override val id: UuidStr,
         override val title: String,
