+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/reminderRepository/reminderRepositoryImpls/ReminderRepositoryImpl.kt
@@ -10,7 +10,7 @@ import com.realityexpander.tasky.agenda_feature.data.repositories.reminderReposi
 import com.realityexpander.tasky.agenda_feature.data.repositories.syncRepository.ISyncRepository
 import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
 import com.realityexpander.tasky.agenda_feature.domain.IReminderRepository
-import com.realityexpander.tasky.agenda_feature.domain.ResultUiText
+import com.realityexpander.tasky.core.presentation.util.ResultUiText
 import com.realityexpander.tasky.core.presentation.util.UiText
 import com.realityexpander.tasky.core.util.ConnectivityObserver.InternetConnectivityObserverImpl.Companion.isInternetReachable
 import com.realityexpander.tasky.core.util.rethrowIfCancellation
