+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/syncRepository/ISyncRepository.kt
@@ -4,7 +4,7 @@ import com.realityexpander.tasky.agenda_feature.common.util.AgendaItemId
 import com.realityexpander.tasky.agenda_feature.data.repositories.syncRepository.local.ModificationTypeForSync
 import com.realityexpander.tasky.agenda_feature.data.repositories.syncRepository.local.entities.SyncItemEntity
 import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
-import com.realityexpander.tasky.agenda_feature.domain.ResultUiText
+import com.realityexpander.tasky.core.presentation.util.ResultUiText
 
 interface ISyncRepository {
     suspend fun addCreatedSyncItem(agendaItem: AgendaItem): ResultUiText<Void>
