+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/domain/IEventRepository.kt
@@ -1,6 +1,7 @@
 package com.realityexpander.tasky.agenda_feature.domain
 
 import com.realityexpander.tasky.agenda_feature.common.util.EventId
+import com.realityexpander.tasky.core.presentation.util.ResultUiText
 import kotlinx.coroutines.flow.Flow
 import java.time.ZonedDateTime
